[feat] - display as wire frame intended
