+++ client/src/components/AdvEntertainment/AdvEntertainment.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import axios from 'axios';
-
+import entertainmentimg from './ferris-wheel.png';
 
 class AdvEntertainment extends React.Component {
   constructor(props) {
@@ -43,6 +43,7 @@ class AdvEntertainment extends React.Component {
     return (
       <div className="adventure-container" onClick={this.handleClick}>
         <h4>Entertainment</h4>
+        <img src={entertainmentimg} alt="no entertainment for you" />
         <h3>{this.state.name}</h3>
         <h6>{this.state.address0}</h6>
         <h6>{this.state.address1}</h6>
