+++ client/src/components/AdvMovies/AdvMovies.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import axios from 'axios';
-
+import movieimg from './video-camera.png'
 
 class AdvMovies extends React.Component {
   constructor(props) {
@@ -37,6 +37,7 @@ class AdvMovies extends React.Component {
     return (
       <div className="adventure-container" onClick={this.handleClick}>
         <h4>Movies</h4>
+        <img src={movieimg} alt="No movie for you" />
         <h3>{this.state.movieName}</h3>
         <h6>{this.state.cinemaName}</h6>
         <img className="images" src={this.state.moviePoster} alt={this.state.name} />
