+++ client/src/components/AdvRandom/AdvRandom.jsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import axios from 'axios';
+import random from './help-round-button.png';
 
 
-class AdvActive extends React.Component {
+class AdvFood extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -19,9 +20,9 @@ class AdvActive extends React.Component {
     const context = this;
     axios({
       method: 'post',
-      url: '/active',
+      url: '/food',
       data: {
-        zipcode: 90045 // NEED TO CHANGE
+        zipcode: 90045  // NEED TO CHANGE TO GRAB INPUT
       }
     }).then(function(res) {
       const item = res.data.businesses[Math.floor(Math.random()*res.data.businesses.length)];
@@ -33,13 +34,13 @@ class AdvActive extends React.Component {
       });
       console.log(item)
     });
-
   }
 
   render() {
     return (
       <div className="adventure-container" onClick={this.handleClick}>
-        <h4>Active</h4>
+        <h4>Food</h4>
+        <img src={random} alt="hahahaha" />
         <h3>{this.state.name}</h3>
         <h6>{this.state.address0}</h6>
         <h6>{this.state.address1}</h6>
@@ -51,4 +52,4 @@ class AdvActive extends React.Component {
 
 
 
-export default AdvActive;
+export default AdvFood;
